+++ src/test/native/index.js
@@ -1,0 +1,3 @@
+var co = require('../../lib/');
+
+module.exports = co;