+++ src/test/withUse.test.js
@@ -1,0 +1,20 @@
+/* jshint expr: true */
+/* global describe */
+
+// shim mz/fs.readFile()
+var pathModule = require('path'),
+    mz = require('mz/fs'),
+    readFile = mz.readFile;
+
+if (!readFile.shimmed) {
+    mz.readFile = function(filename, encoding) {
+        return readFile(pathModule.join(__dirname, 'native', filename), encoding);
+    };
+    mz.readFile.shimmed = true;
+}
+
+// run co tests
+
+describe('With use(bluebird)', function() {
+    /* insert requires */
+});