+++ src/test/withUse/index.js
@@ -1,0 +1,4 @@
+var co = require('../../lib/'),
+    Bluebird = require('bluebird');
+
+module.exports = co.use(Bluebird);