+++ changelog.md
@@ -10,3 +10,4 @@
 * co-use tests use generators
 * `npm run build` command
 * Do not run jshint on built tests
+* Build loads newest co and creates tests files for co tests